requirements.txt
fastapi
uvicorn
motor
pydantic
langchain
langchain-google-genai
python-dotenv
pytest
httpx
